requirements.txt
streamlit~=1.48.0
google-generativeai
python-dotenv~=1.1.1
pip~=25.2
attrs~=25.3.0
distro~=1.9.0
wheel~=0.41.2
pillow~=11.3.0
tornado~=6.5.1
Jinja2~=3.1.6
setuptools~=78.1.1
packaging~=25.0
pyparsing~=3.2.3
typing_extensions~=4.14.1